+++ changes.md
@@ -2,6 +2,7 @@
 
 ## 3.6.0
 * Update to cadqueryhelper version 2.0.5
+* Updated cadqueryhelper import paths that reference grid components.
 
 ## 3.5.6
 * Update to cadqueryhelper version 1.7.3